+++ client/src/components/dashboard/PostCard.js
@@ -3,10 +3,10 @@ import { Button, Card, Icon, Label, Image } from 'semantic-ui-react';
 import { Link } from 'react-router-dom';
 import moment from 'moment';
 
-import { AuthContext } from '../context/auth';
+import { AuthContext } from '../../context/auth';
 import LikeButton from './LikeButton';
 import DeleteButton from './DeleteButton';
-import MyPopup from '../util/MyPopup';
+import MyPopup from '../../util/MyPopup';
 
 function PostCard({
   post: { body, createdAt, id, username, likeCount, commentCount, likes }
