+++ client/src/components/landing-page/Footer/FooterElements.js
