+++ client/src/components/landing-page/Footer/index.js
