+++ client/src/components/landing-page/HeroSection/index.js
@@ -1,5 +1,5 @@
 import React, { useState } from 'react'
-import Video from '../../assets/videos/leaf-lake.mp4'
+import Video from '../../../assets/videos/leaf-lake.mp4'
 import { HeroContainer, HeroBg, VideoBg, HeroContent, HeroH1, HeroP, HeroBtnWrapper, ArrowForward, ArrowRight } from './HeroElements'
 import { Button } from '../ButtonElement'
