+++ client/src/components/landing-page/InfoSection/Data.js
@@ -8,7 +8,7 @@ export const homeObjOne = {
     description: 'Get access to our exclusive online platform that lets you interact with other shower philosophers without charging you anything for it.',
     buttonLabel: 'Get Started',
     imgStart: false,
-    img: require('../../assets/images/ideas.svg').default,
+    img: require('../../../assets/images/ideas.svg').default,
     alt: 'Car',
     dark: true,
     primary: true,
@@ -25,7 +25,7 @@ export const homeObjTwo = {
     description: 'Get access to our exclusive online platform that lets you interact with other shower philosophers without charging you anything for it.',
     buttonLabel: 'Learn More',
     imgStart: true,
-    img: require('../../assets/images/social.svg').default,
+    img: require('../../../assets/images/social.svg').default,
     alt: 'Social Clarity',
     dark: false,
     primary: false,
@@ -42,7 +42,7 @@ export const homeObjThree = {
     description: 'Get access to our exclusive online platform that lets you interact with other shower philosophers without charging you anything for it.',
     buttonLabel: 'Start Now',
     imgStart: false,
-    img: require('../../assets/images/meeting.svg').default,
+    img: require('../../../assets/images/meeting.svg').default,
     alt: 'Discussions',
     dark: false,
     primary: false,
