+++ client/src/components/landing-page/InfoSection/index.js
