+++ client/src/components/landing-page/Navbar/NavBarElements.js
