+++ client/src/pages/Dashboard.js
@@ -1,0 +1,46 @@
+import React, { useContext } from 'react';
+import { useQuery } from '@apollo/react-hooks';
+import { Grid, Transition } from 'semantic-ui-react';
+
+import { AuthContext } from '../context/auth';
+import PostCard from '../components/dashboard/PostCard';
+import PostForm from '../components/dashboard/PostForm';
+import { FETCH_POSTS_QUERY } from '../util/graphql';
+
+function Dashboard() {
+  const { user } = useContext(AuthContext);
+  let posts = "";
+  const { loading, data } = useQuery(FETCH_POSTS_QUERY);
+  if (data) {
+    posts = data.getPosts;
+  }
+
+  return (
+    <Grid columns={3}>
+      <Grid.Row className="page-title">
+        <h1>Recent Posts</h1>
+      </Grid.Row>
+      <Grid.Row>
+        {user && (
+          <Grid.Column>
+            <PostForm />
+          </Grid.Column>
+        )}
+        {loading ? (
+          <h1>Loading posts..</h1>
+        ) : (
+          <Transition.Group>
+            {posts &&
+              posts.map((post) => (
+                <Grid.Column key={post.id} style={{ marginBottom: 20 }}>
+                  <PostCard post={post} />
+                </Grid.Column>
+              ))}
+          </Transition.Group>
+        )}
+      </Grid.Row>
+    </Grid>
+  );
+}
+
+export default Dashboard;
