+++ client/src/pages/Home.js
@@ -1,46 +1,32 @@
-import React, { useContext } from 'react';
-import { useQuery } from '@apollo/react-hooks';
-import { Grid, Transition } from 'semantic-ui-react';
+import React, { useState } from 'react'
 
-import { AuthContext } from '../context/auth';
-import PostCard from '../components/PostCard';
-import PostForm from '../components/PostForm';
-import { FETCH_POSTS_QUERY } from '../util/graphql';
+import Navbar from '../components/landing-page/Navbar';
+import Sidebar from '../components/landing-page/Sidebar'
+import HeroSection from '../components/landing-page/HeroSection';
+import InfoSection from '../components/landing-page/InfoSection';
+import { homeObjOne, homeObjThree, homeObjTwo } from '../components/landing-page/InfoSection/Data';
+import Services from '../components/landing-page/Services'
+import Footer from '../components/landing-page/Footer';
 
-function Home() {
-  const { user } = useContext(AuthContext);
-  let posts = "";
-  const { loading, data } = useQuery(FETCH_POSTS_QUERY);
-  if (data) {
-    posts = data.getPosts;
-  }
+const Home = () => {
+    const [isOpen, setIsOpen] = useState(false)
 
-  return (
-    <Grid columns={3}>
-      <Grid.Row className="page-title">
-        <h1>Recent Posts</h1>
-      </Grid.Row>
-      <Grid.Row>
-        {user && (
-          <Grid.Column>
-            <PostForm />
-          </Grid.Column>
-        )}
-        {loading ? (
-          <h1>Loading posts..</h1>
-        ) : (
-          <Transition.Group>
-            {posts &&
-              posts.map((post) => (
-                <Grid.Column key={post.id} style={{ marginBottom: 20 }}>
-                  <PostCard post={post} />
-                </Grid.Column>
-              ))}
-          </Transition.Group>
-        )}
-      </Grid.Row>
-    </Grid>
-  );
+    const toggle = () => {
+        setIsOpen(!isOpen)
+    }
+
+    return (
+        <div>
+            <Sidebar isOpen={isOpen} toggle={toggle}/>
+            <Navbar toggle={toggle}/>
+            <HeroSection />
+            <InfoSection {...homeObjOne} />
+            <InfoSection {...homeObjTwo} />
+            <Services />
+            <InfoSection {...homeObjThree} />
+            <Footer />
+        </div>
+    )
 }
 
 export default Home;
