+++ client/src/util/ApolloProvider.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import App from './App';
+import App from '../App';
 import ApolloClient from 'apollo-client';
 import { InMemoryCache } from 'apollo-cache-inmemory';
 import { createHttpLink } from 'apollo-link-http';
